+++ Cargo.toml
@@ -24,7 +24,7 @@ daemonize = "0.4.1"
 lz4 = "1.23.2"
 regex = "1"
 lazy_static = "1.4.0"
-
+nom = "6"
 
 # We'll use cloudflare's zlib as it is the fastest game in town
 #flate2 = "1.0" #minz library
