+++ src/feature_transform_parser.rs
@@ -1,3 +1,6 @@
+//#[macro_use]
+//extern crate nom;
+
 use crate::model_instance;
 use crate::parser;
 use crate::vwmap;
@@ -10,6 +13,8 @@ use serde::{Serialize,Deserialize};
 
 use regex::Regex;
 use lazy_static::lazy_static; 
+
+
 
 lazy_static! {
     static ref NAMESPACE_TRANSFORM_REGEX: Regex = Regex::new(r"^(.)=(\w+)\((.)\)$").unwrap();
@@ -109,47 +114,6 @@ pub fn get_namespace_id(transform_namespaces: &TransformNamespaces, vw: &vwmap::
        }
    };   
 }
-
-
-
-
-#[inline(always)]
-fn emit_u32(hash_data:u32, hash_value:f32, namespace_seed: u32) -> (u32, f32) {
-    (murmur3::hash32_with_seed(hash_data.to_le_bytes(), namespace_seed) & parser::MASK31, hash_value)
-}                                                         
-
-
-pub fn transformed_feature<'a>(record_buffer: &[u32], mi: &model_instance::ModelInstance, feature_index_offset: u32) -> Vec<(u32, f32)> {
-    // This is FAAAR from optimized
-    let mut output:Vec<(u32, f32)> = Vec::new();
-    let feature_index_offset = feature_index_offset & !TRANSFORM_NAMESPACE_MARK; // remove transform namespace mark
-    //println!("Fi {}", feature_index_offset);
-    let transform_namespace = &mi.transform_namespaces.v[feature_index_offset as usize];
-    if transform_namespace.function == TransformFunction::Sqrt {
-        feature_reader_float_namespace!(record_buffer, transform_namespace.from_namespace_index, hash_data, hash_value, float_value, {
-            let transformed_float = float_value.sqrt();
-            let transformed_int = transformed_float as u32;
-            output.push(emit_u32(transformed_int, hash_value, 0));
-            //println!("Input hash value {}, float value {}", hash_value, float_value);
-            //println!("Sqrt: {}, sqrt_int {}", transformed_float, transformed_int);
-            //println!("Output hash data {}, hash_value {}", output.last().unwrap().0, output.last().unwrap().1);
-              
-        });
-    }
-
-    output
-}
-
-
-
-
-
-
-
-
-
-
-
