+++ src/main.rs
@@ -33,7 +33,8 @@ mod block_lr;
 mod block_loss_functions;
 mod block_helpers;
 mod multithread_helpers;
-mod feature_transform;
+mod feature_transform_parser;
+mod feature_transform_executor;
 
 fn main() {
     match main2() {
